+++ src/write-functions/write-code.js
@@ -1,5 +1,6 @@
 let { sep } = require('path')
 let { writeFileSync } = require('fs')
+let { aliases } = require('lambda-runtimes')
 let http = require('./templates/http')
 let events = require('./templates/events')
 let queues = require('./templates/queues')
@@ -12,13 +13,15 @@ module.exports = function writeCode (params) {
   let handler = handlerFile.replace(process.cwd(), '')
   if (handler[0] === sep) handler = handler.substr(1)
 
-  let types = { http, events, queues, ws, scheduled, streams }
-  let run = 'node'
+  runtime = aliases[runtime] ? aliases[runtime] : runtime
+  let run
   if (runtime.startsWith('deno'))   run = 'deno'
   if (runtime.startsWith('node'))   run = 'node'
   if (runtime.startsWith('python')) run = 'python'
   if (runtime.startsWith('ruby'))   run = 'ruby'
+  if (!run) throw ReferenceError(`Valid runtime not found: ${params.runtime}`)
 
+  let types = { http, events, queues, ws, scheduled, streams }
   if (!body) body = pragma === 'http'
     ? types[pragma][run](handler)
     : types[pragma][run]