+++ test/unit/src/bootstrap/_arc-package-test.js
@@ -21,13 +21,14 @@ let fsStub = {
 let arcPackage = proxyquire('../../../../src/bootstrap/_arc-package', {
   fs: fsStub
 })
-let options = []
 let foo = 'foo'
 let name = foo
 let folder = __dirname
+let argv = process.argv
 let reset = () => {
   destination = ''
   written = ''
+  process.argv = argv
 }
 
 test('Set up env', t => {
@@ -37,18 +38,21 @@ test('Set up env', t => {
 
 test('Package writer bails when Architect is called from a global install', t => {
   t.plan(2)
-  let options = [ '/usr/local/bin/node', '/usr/local/bin/arc', 'create' ]
-  let result = arcPackage({ options })
+  process.argv = [ '/usr/local/bin/node', '/usr/local/bin/arc', 'create' ]
+  let result = arcPackage({})
   t.notOk(result, 'Invocation from global install should opt out of Arc installation')
+  reset()
 
-  options = [ '/usr/local/bin/node', '/usr/local/bin/arc', 'init' ]
-  result = arcPackage({ options })
+  process.argv = [ '/usr/local/bin/node', '/usr/local/bin/arc', 'init' ]
+  result = arcPackage({})
   t.notOk(result, 'Invocation from global install should opt out of Arc installation')
+  reset()
 })
 
 test('Package writer writes a package file when none exists', t => {
   t.plan(3)
-  let result = arcPackage({ options, name, folder })
+  process.argv = []
+  let result = arcPackage({ name, folder })
   t.ok(result, 'Missing package file found should opt into Arc installation')
   t.equal(destination, join(__dirname, 'package.json'), 'Wrote package.json to specified folder')
   t.equal(JSON.parse(written)['name'], foo, 'package.json uses specified app name')
@@ -57,15 +61,19 @@ test('Package writer writes a package file when none exists', t => {
 
 test(`Package writer determines whether to install Arc based on an existing package.json`, t => {
   t.plan(3)
+  process.argv = []
   exists = true
-  let result = arcPackage({ options, name, folder })
+  let result = arcPackage({ name, folder })
   t.notOk(result, 'Found package file with Arc as dep should opt out of Arc installation')
+  reset()
 
   pkg = arcAsDevDep
-  result = arcPackage({ options, name, folder })
+  result = arcPackage({ name, folder })
   t.notOk(result, 'Found package file with Arc as dev dep should opt out of Arc installation')
+  reset()
 
   pkg = {}
-  result = arcPackage({ options, name, folder })
+  result = arcPackage({ name, folder })
   t.ok(result, 'Found package file with no Arc installed, so opt into Arc installation')
+  reset()
 })